backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
uvloop>=0.19.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.31
asyncpg>=0.29.0
alembic>=1.13.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
python-multipart>=0.0.9
docker>=7.1.0
chromadb>=1.0.0
pypdf>=4.3.0
python-docx>=1.1.2
redis>=5.0.7
